List quarters (#9)
* Create quarters controller with index and search actions

* Add swagger documentation

+++ app/controllers/swagger/models/city.rb
@@ -11,7 +11,7 @@ module Swagger
           key :required, %i[id name alpha_2_code]
 
           property :id do
-            key :type, :int64
+            key :type, :integer
             key :example, 1
           end
